+++ index.html
@@ -235,7 +235,7 @@
     </template>
 
   <!-- Подключение JS -->
-  <script src="/js/main.js"></script>
+  <script type='module' src="/js/main.js"></script>
   <script src="/js/functions.js"></script>
 </body>
 </html>
