+++ js/modules/dataset.js
@@ -1,0 +1,46 @@
+
+const NAMES = [
+  'Антон',
+  'Александра',
+  'Артём',
+  'София',
+  'Матвей',
+  'Анна',
+  'Иван',
+  'Полина',
+  'Тимофей',
+  'Ксения',
+  'Кирилл',
+  'Дарья',
+  'Арсений',
+  'Елизавета',
+  'Андрей',
+  'Злата',
+  'Сергей',
+  'Виктория',
+  'Елисей',
+  'Мария',
+  'Олег',
+  'Анастасия',
+  'Макар',
+  'Алиса'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Личное фото',
+  'Рабочее фото',
+  'Семейное фото',
+  'Детское фото'
+];
+
+export { NAMES, MESSAGES, DESCRIPTIONS };
